common util
